Move the thread index to here.
